+++ mozilla/accessible/src/base/nsAccessiblePivot.cpp
@@ -14,7 +14,7 @@ using namespace mozilla::a11y;
 
 
 /**
- * An object that stores a given traversal rule during the pivot movement.
+ * An object that stores a given traversal rule during 
  */
 class RuleCache
 {
@@ -325,7 +325,7 @@ nsAccessiblePivot::MoveNextByText(TextBoundaryType aBoundary, bool* aResult)
 
     // If there's no more text on the current node, try to find the next text
     // node; if there isn't one, bail out.
-    if (tempEnd == static_cast<int32_t>(text->CharacterCount())) {
+    if (tempEnd == text->CharacterCount()) {
       if (tempPosition == root)
         return NS_OK;
 
@@ -389,7 +389,7 @@ nsAccessiblePivot::MoveNextByText(TextBoundaryType aBoundary, bool* aResult)
     // instead want to traverse into it. So restart the movement with
     // the child as the starting point.
     if (childAtOffset && nsAccUtils::IsEmbeddedObject(childAtOffset) &&
-        tempStart == static_cast<int32_t>(childAtOffset->StartOffset())) {
+        tempStart == childAtOffset->StartOffset()) {
       tempPosition = childAtOffset;
       tempStart = tempEnd = -1;
       continue;
@@ -517,7 +517,7 @@ nsAccessiblePivot::MovePreviousByText(TextBoundaryType aBoundary, bool* aResult)
     // instead want to traverse into it. So restart the movement with
     // the child as the starting point.
     if (childAtOffset && nsAccUtils::IsEmbeddedObject(childAtOffset) &&
-        tempEnd == static_cast<int32_t>(childAtOffset->EndOffset())) {
+        tempEnd == childAtOffset->EndOffset()) {
       tempPosition = childAtOffset;
       tempStart = tempEnd = childAtOffset->AsHyperText()->CharacterCount();
       continue;
