+++ mozilla/dom/xslt/base/txExpandedNameMap.cpp
@@ -26,7 +26,7 @@ class txMapItemComparator
 nsresult txExpandedNameMap_base::addItem(const txExpandedName& aKey,
                                          void* aValue)
 {
-    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         return NS_ERROR_XSLT_ALREADY_SET;
     }
@@ -53,10 +53,11 @@ nsresult txExpandedNameMap_base::setItem(const txExpandedName& aKey,
                                          void** aOldValue)
 {
     *aOldValue = nullptr;
-    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         *aOldValue = mItems[pos].mValue;
         mItems[pos].mValue = aValue;
+        
         return NS_OK;
     }
 
@@ -77,7 +78,7 @@ nsresult txExpandedNameMap_base::setItem(const txExpandedName& aKey,
  */
 void* txExpandedNameMap_base::getItem(const txExpandedName& aKey) const
 {
-    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         return mItems[pos].mValue;
     }
@@ -94,7 +95,7 @@ void* txExpandedNameMap_base::getItem(const txExpandedName& aKey) const
 void* txExpandedNameMap_base::removeItem(const txExpandedName& aKey)
 {
     void* value = nullptr;
-    size_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
+    uint32_t pos = mItems.IndexOf(aKey, 0, txMapItemComparator());
     if (pos != mItems.NoIndex) {
         value = mItems[pos].mValue;
         mItems.RemoveElementAt(pos);
