+++ mozilla/xpcom/glue/nsTArrayForwardDeclare.h
@@ -19,7 +19,7 @@
 // template parameters.
 //
 
-#include <stddef.h>
+#include <stdint.h>
 
 template<class E>
 class nsTArray;
@@ -27,10 +27,10 @@ class nsTArray;
 template<class E>
 class FallibleTArray;
 
-template<class E, size_t N>
+template<class E, uint32_t N>
 class nsAutoTArray;
 
-template<class E, size_t N>
+template<class E, uint32_t N>
 class AutoFallibleTArray;
 
 #define InfallibleTArray nsTArray
